requirements.txt
kafka-python
cassandra-driver
pandas